added checks in services of info table

+++ src/main/java/kg/attractor/job_search/service/impl/EducationInfoServiceImpl.java
@@ -2,6 +2,7 @@ package kg.attractor.job_search.service.impl;
 
 import kg.attractor.job_search.dao.EducationInfoDao;
 import kg.attractor.job_search.dto.EducationInfoDto;
+import kg.attractor.job_search.exception.ResumeNotFoundException;
 import kg.attractor.job_search.model.EducationInfo;
 import kg.attractor.job_search.service.EducationInfoService;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +21,8 @@ public class EducationInfoServiceImpl implements EducationInfoService {
     @Override
     @Transactional
     public void createEducationInfo(List<EducationInfoDto> educationInfoDtoList, Integer resumeId) {
-        if (educationInfoDtoList != null || !educationInfoDtoList.isEmpty()) {
+        if (educationInfoDtoList != null) {
+            if (!educationInfoDtoList.isEmpty()) {
             educationInfoDtoList.forEach(educationInfoDto -> {
                 EducationInfo educationInfo = new EducationInfo();
                 educationInfo.setResumeId(resumeId);
@@ -31,16 +33,23 @@ public class EducationInfoServiceImpl implements EducationInfoService {
                 educationInfo.setStartDate(educationInfoDto.getStartDate());
                 educationInfoDao.createEducationInfo(educationInfo);
             });
+            }
         }
     }
 
     @Override
     public void deleteEducationInfoByResumeId(Integer resumeId) {
+        if (educationInfoDao.findById(resumeId).isEmpty()) {
+            throw new ResumeNotFoundException("Resume with id " + resumeId + " not found");
+        }
         educationInfoDao.deleteEducationInfoByResumeId(resumeId);
     }
 
     @Override
     public List<EducationInfoDto> getEducationInfoByResumeId(Integer resumeId) {
+        if (educationInfoDao.findById(resumeId).isEmpty()) {
+            throw new ResumeNotFoundException("Resume with id " + resumeId + " not found");
+        }
         List<EducationInfo> educationInfoList = educationInfoDao.findEducationInfoByResumeId(resumeId);
 
         return educationInfoList.stream()
