+++ src/main/java/kg/attractor/job_search/service/impl/WorkExperienceInfoServiceImpl.java
@@ -2,6 +2,7 @@ package kg.attractor.job_search.service.impl;
 
 import kg.attractor.job_search.dao.WorkExperienceInfoDao;
 import kg.attractor.job_search.dto.WorkExperienceInfoDto;
+import kg.attractor.job_search.exception.ResumeNotFoundException;
 import kg.attractor.job_search.model.WorkExperienceInfo;
 import kg.attractor.job_search.service.WorkExperienceInfoService;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +21,8 @@ public class WorkExperienceInfoServiceImpl implements WorkExperienceInfoService 
     @Transactional
     @Override
     public void createWorkExperienceInfo(List<WorkExperienceInfoDto> workExperienceInfoDtoList, Integer resumeId) {
-        if (workExperienceInfoDtoList != null && !workExperienceInfoDtoList.isEmpty()) {
+        if (workExperienceInfoDtoList != null) {
+            if (!workExperienceInfoDtoList.isEmpty()) {
             workExperienceInfoDtoList.forEach(workExperienceInfoDto -> {
                 WorkExperienceInfo workExperienceInfo = new WorkExperienceInfo();
                 workExperienceInfo.setResumeId(resumeId);
@@ -30,16 +32,23 @@ public class WorkExperienceInfoServiceImpl implements WorkExperienceInfoService 
                 workExperienceInfo.setResponsibilities(workExperienceInfoDto.getResponsibilities());
                 workExperienceInfoDao.createWorkExperienceInfo(workExperienceInfo);
             });
+            }
         }
     }
 
     @Override
     public void deleteWorkExperienceInfoByResumeId(Integer resumeId) {
+        if (workExperienceInfoDao.findById(resumeId).isEmpty()) {
+            throw new ResumeNotFoundException("Resume with id " + resumeId + " not found");
+        }
         workExperienceInfoDao.deleteWorkExperienceInfoByResumeId(resumeId);
     }
 
     @Override
     public List<WorkExperienceInfoDto> getWorkExperienceInfoByResumeId(Integer resumeId) {
+        if (workExperienceInfoDao.findById(resumeId).isEmpty()) {
+            throw new ResumeNotFoundException("Resume with id " + resumeId + " not found");
+        }
         List<WorkExperienceInfo> workExperienceInfoList = workExperienceInfoDao.findWorkExperienceInfoByResumeId(resumeId);
 
         return workExperienceInfoList.stream()
